requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.19.0
python-dotenv==0.19.0
cachetools==4.2.4
pydantic==1.8.2
pytest==6.2.5
pytest-asyncio==0.15.1
